+++ packages/vue-component-library/src/lib-components/BlockPostSmall.vue
@@ -1,58 +1,47 @@
-<script
-  lang="ts"
-  setup
->
-import type { PropType } from 'vue'
+<script lang="ts" setup>
 import { computed } from 'vue'
+import type { PropType } from 'vue'
+import getSectionName from '@/utils/getSectionName'
 
-// TYPESCRIPT
+// Child components
+import SmartLink from '@/lib-components/SmartLink.vue'
+
+// TYPES
 import type { MediaItemType } from '@/types/types'
 
-// UTILITY FUNCTION
-import getSectionName from '@/utils/getSectionName'
+// UTILS
+
+// THEME
+import { useTheme } from '@/composables/useTheme'
 
 // COMPONENTS
 import ResponsiveImage from '@/lib-components/ResponsiveImage.vue'
 
-const props = defineProps({
-  image: {
-    type: Object as PropType<MediaItemType>,
-    default: () => { },
-  },
-  categoryName: {
-    type: String,
-    default: '',
-  },
-  title: {
-    type: String,
-    default: '',
-  },
-  author: {
-    type: String,
-    default: '',
-  },
+const { to, image } = defineProps({
   to: {
     type: String,
-    default: '',
+    required: true,
+  },
+  image: {
+    type: Object as PropType<MediaItemType>,
+    default: () => ({}),
   },
 })
 
-const sectionName = computed(() => {
-  return getSectionName(props.to)
-})
+// THEME & SECTION COLOR
+const theme = useTheme()
+const sectionName = computed(() => getSectionName(to))
 
 const classes = computed(() => {
-  return ['block-post-small', `color-${sectionName.value}`]
-})
-
-const parsedAuthor = computed(() => {
-  return `By ${props.author}`
+  return theme?.value === 'ftva'
+    ? ['block-post-small', 'ftva']
+    : ['block-post-small', `color-${sectionName.value}`]
 })
 </script>
 
 <template>
   <div>
-    <smart-link
+    <SmartLink
       :to="to"
       :class="classes"
     >
@@ -62,101 +51,25 @@ const parsedAuthor = computed(() => {
         :aspect-ratio="100"
         class="image"
       />
+
       <div class="meta">
-        <div
-          class="category"
-          v-html="categoryName"
-        />
-        <h3
-          class="title"
-          v-html="title"
-        />
-        <div
-          class="author"
-          v-html="parsedAuthor"
-        />
+        <div class="category">
+          <slot name="category" />
+        </div>
+
+        <h3 class="title">
+          <slot name="title" />
+        </h3>
+
+        <div class="author">
+          <slot name="author" />
+        </div>
       </div>
-    </smart-link>
+    </SmartLink>
   </div>
 </template>
 
-<style
-  lang="scss"
-  scoped
->
-.block-post-small {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  justify-content: space-between;
-  align-content: space-between;
-  align-items: center;
-
-  width: 320px;
-  min-height: 150px;
-  padding: 25px;
-  border-radius: var(--rounded-slightly-all);
-  transition-property: background-color, box-shadow;
-  transition-duration: 400ms;
-  transition-timing-function: ease-in-out;
-
-  .image {
-    width: 100px;
-    flex-shrink: 0;
-  }
-
-  .meta {
-    max-width: 150px;
-    margin-left: 20px;
-  }
-
-  .category {
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 100%;
-    letter-spacing: 0.06em;
-    text-transform: uppercase;
-    color: var(--color-primary-blue-05);
-  }
-
-  .title {
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 120%;
-    letter-spacing: 0.01em;
-    color: var(--color-primary-blue-03);
-    margin: 15px 0 0 0;
-  }
-
-  .author {
-    font-weight: 300;
-    font-size: 14px;
-    line-height: 100%;
-    color: var(color-grey-01);
-    margin-top: 10px;
-  }
-
-  // Themes
-  --color-theme: var(--color-primary-blue-02);
-
-  &.color-visit {
-    --color-theme: var(--color-visit-fushia-01);
-  }
-
-  &.color-help {
-    --color-theme: var(--color-help-green-01);
-  }
-
-  &.color-about {
-    --color-theme: var(--color-about-purple-01);
-  }
-
-  // Hovers
-  @media #{$has-hover} {
-    &:hover {
-      background-color: var(--color-theme);
-      box-shadow: 0px 10px 17px rgba(0, 0, 0, 0.04);
-    }
-  }
-}
+<style lang="scss" scoped>
+@import "@/styles/default/_block-post-small.scss";
+@import "@/styles/ftva/_block-post-small.scss";
 </style>