+++ packages/vue-component-library/src/stories/SectionPostSmall.stories.js
@@ -1,3 +1,5 @@
+import { computed } from 'vue'
+
 // Import mock api data
 import * as API from '@/stories/mock-api.json'
 import SectionPostSmall from '@/lib-components/SectionPostSmall'
@@ -18,7 +20,7 @@ const mock = [
   },
   {
     image: API.image,
-    to: ' ',
+    to: '/visit/fizz',
     categoryName: 'wild pets',
     author: 'Dexter Diamond',
     title: 'Introducing Mammals to Young Naturalists',
@@ -50,6 +52,50 @@ export function Default() {
       <section-post-small
         :items="items"
         to="/visit/foo/bar/"
+      />
+  `,
+  }
+}
+
+const mockFtva = [
+  {
+    image: API.image,
+    to: '/collection-policy',
+    title: 'Collection Policy',
+  },
+  {
+    image: API.image,
+    to: '/media-licensing',
+    title: 'Media Licensing',
+  },
+  {
+    image: API.image,
+    to: 'https://www.cinema.ucla.edu/donating-materials',
+    title: 'Donating Materials',
+  },
+  {
+    image: API.image,
+    to: '/media-loans',
+    title: 'Media Loans',
+  }
+]
+
+export function Ftva() {
+  return {
+    data() {
+      return { items: mockFtva }
+    },
+    provide() {
+      return {
+        theme: computed(() => 'ftva'),
+      }
+    },
+    components: { SectionPostSmall },
+    template: `
+      <SectionPostSmall
+        :items="items"
+        sectionTitle="About Our Collections"
+        sectionSummary="<p>With over 350,00 motion pictures and 170,000 television programs, and 27 million feet of newsreel footage, the <em>UCLA Film & Television Archive</em> is the largest university-held collection of motion pictures and broadcast programming.</p><p>For more information on our collections or to arrange research viewing please contact the <strong>Archive Research and Study Center</strong> (ARSC).</p>"
       />
   `,
   }